move all blocks to mysite.blocks
