+++ mysite/blocks.py
@@ -1,14 +1,96 @@
 from django.db import models
 from django import forms
 
-from wagtail.wagtailcore.models import Page
-from wagtail.wagtailcore.fields import StreamField, RichTextField
 from wagtail.wagtailcore import blocks
 from wagtail.wagtailembeds.blocks import EmbedBlock
 from wagtail.wagtailimages.blocks import ImageChooserBlock
 from wagtail.wagtailembeds.blocks import EmbedBlock
 from wagtail.contrib.table_block.blocks import TableBlock
-from home.blocks import IntegerBlock
+from wagtail.wagtailcore.blocks import IntegerBlock
+
+class CustomImageBlock(blocks.StructBlock):
+	image = ImageChooserBlock(icon="image", required=True)
+	align = blocks.ChoiceBlock(choices=[
+		('left', 'Left'),
+		('right', 'Right'),
+		('full-width', 'Full Width')
+	], required=True)
+	width = blocks.ChoiceBlock([
+		('initial', 'Auto'),
+		('60%', '60%'),
+		('50%', '50%'),
+		('33.333%', '33%'),
+		('25%', '25%')
+	], default="initial", required=True)
+
+	class Meta:
+		template = 'blocks/image_block.html'
+
+class ButtonBlock(blocks.StructBlock):
+	button_text = blocks.CharBlock(required=True, max_length=50)
+	button_link = blocks.URLBlock(required=True, default="https://www.")
+	alignment = blocks.ChoiceBlock(choices=[
+		('left-aligned', 'Left'),
+		('center-aligned', 'Center')
+	])
+
+	class Meta:
+		template = 'blocks/button.html'
+		icon = 'radio-full'
+		label = 'Button'
+
+class IframeBlock(blocks.StructBlock):
+	source_url = blocks.URLBlock(required=True)
+	width = IntegerBlock(max_value=1050, help_text="The maximum possible iframe width is 1050")
+	height = IntegerBlock()
+
+	class Meta:
+		template = 'blocks/iframe.html'
+		icon = 'form'
+		label = 'Iframe'
+		help_text= "Specifiy maximum width and height dimensions for the iframe. On smaller screens, width-to-height ratio will be preserved."
+
+class DatavizBlock(blocks.StructBlock):
+	title = blocks.CharBlock(required=False)
+	subheading = blocks.RichTextBlock(required=False)
+	max_width = IntegerBlock()
+	show_chart_buttons = blocks.BooleanBlock(default=False, required=False)
+	container_id = blocks.CharBlock(required=True)
+
+	class Meta:
+		template = 'blocks/dataviz.html'
+		icon = 'site'
+		label = 'Dataviz'
+
+class TwoColumnBlock(blocks.StructBlock):
+    left_column = blocks.RichTextBlock()
+    right_column = blocks.RichTextBlock()
+
+    class Meta:
+        template = 'blocks/two-column.html'
+
+class IntegerChoiceBlock(blocks.ChoiceBlock):
+    choices = (
+        ('1', '1'),
+        ('2', '2'),
+        ('3', '3'),
+        ('4', '4'),
+        ('5','5'),
+        ('6', '6')
+    )
+
+class PersonBlock(blocks.StructBlock):
+    name = blocks.TextBlock(required=True)
+    title = blocks.TextBlock(required=False, max_length=125, help_text="125 character limit")
+    description = blocks.RichTextBlock(required=False)
+    image = ImageChooserBlock(icon='image', required=False)
+    twitter = blocks.URLBlock(required=False)
+
+class PeopleBlock(blocks.StreamBlock):
+    person = PersonBlock();
+
+    class Meta:
+        template = 'blocks/people.html'
 
 class GoogleMapBlock(blocks.StructBlock):
     use_page_address = blocks.BooleanBlock(default=False, required=False, help_text="If selected, map will use the address already defined for this page, if applicable. For most posts besides events, this should be left unchecked and the form below should be completed.")
@@ -19,3 +101,60 @@ class GoogleMapBlock(blocks.StructBlock):
 
     class Meta:
         template = 'blocks/google_map.html'
+
+class SessionTypesBlock(blocks.ChoiceBlock):
+    choices = (
+        ('panel', 'Panel'),
+        ('lecture', 'Lecture'),
+        ('break', 'Break'),
+        ('meal', 'Meal'),
+        ('reception','Reception'),
+        ('registration', 'Registration')
+    )
+
+class SessionSpeakerBlock(blocks.StructBlock):
+    name = blocks.TextBlock(required=True)
+    title = blocks.TextBlock(required=False)
+
+class SessionBlock(blocks.StructBlock):
+    name = blocks.TextBlock()
+    session_type = SessionTypesBlock()
+    description = blocks.RichTextBlock(required=False)
+    start_time = blocks.TimeBlock(required=False)
+    end_time = blocks.TimeBlock(required=False)
+    speakers = blocks.StreamBlock([
+        ('speaker', SessionSpeakerBlock())
+    ])
+    archived_video_link = blocks.URLBlock(help_text="Enter youtube link after conference", required=False)
+
+class SessionDayBlock(blocks.StructBlock):
+    day = IntegerChoiceBlock(help_text="What day of the conference is this session on?", required=False)
+    start_time = blocks.TimeBlock(required=False)
+    end_time = blocks.TimeBlock(required=False)
+    sessions = blocks.StreamBlock([
+        ('session', SessionBlock())
+    ])
+
+class SessionsBlock(blocks.StreamBlock):
+    days = SessionDayBlock()
+
+    class Meta:
+        template = 'blocks/schedule.html'
+
+class BodyBlock(blocks.StreamBlock):
+    introduction = blocks.RichTextBlock()
+    heading = blocks.CharBlock(classname='full title')
+    paragraph = blocks.RichTextBlock()
+    inline_image = CustomImageBlock(icon='image')
+    video = EmbedBlock(icon='media')
+    table = TableBlock()
+    button = ButtonBlock()
+    iframe = IframeBlock()
+    dataviz = DatavizBlock()
+    google_map = GoogleMapBlock()
+    image = ImageChooserBlock(template='blocks/image_block.html', help_text='Legacy option. Consider using Inline Image instead.')
+
+
+class EventBodyBlock(BodyBlock):
+    schedule = SessionsBlock()
+    speakers = PeopleBlock()
