+++ mysite/templates/blocks/image_block.html
