+++ mysite/templates/blocks/people.html
