+++ mysite/templates/blocks/two-column.html
